:recycle: Refactored type coloring logic & adjusted some styling
