+++ src/features/utilities.ts
@@ -1,5 +1,18 @@
-import { ActionCreatorWithPayload } from "@reduxjs/toolkit";
 import { Dispatch } from "react";
+import { ActionCreatorWithPayload, PayloadAction } from "@reduxjs/toolkit";
+import { SliceStatus } from "../globals";
+
+export const statusHandlerReducer = {
+  initialize: (state: any, action: PayloadAction) => {
+    state.status.state = SliceStatus.LOADING;
+  },
+  error: (state: any, action: PayloadAction) => {
+    state.status.state = SliceStatus.ERROR;
+  },
+  success: (state: any, action: PayloadAction) => {
+    state.status.state = SliceStatus.SUCCESS;
+  },
+};
 
 type StatusHandler = {
   initialize: ActionCreatorWithPayload<any, string>;
