+++ src/utils/leftPad.ts
@@ -1,0 +1,7 @@
+export const leftPad = (number: number, targetLength: number) => {
+  let output = number + "";
+  while (output.length < targetLength) {
+    output = "0" + output;
+  }
+  return output;
+};
